+++ R/xlsx_sheet_names.R
@@ -6,6 +6,9 @@
 #' is opened with a spreadsheet application like Excel or LibreOffice.
 #'
 #' @param path Path to the xlsx file.
+#' @param check_filetype Logical. Whether to check that the filetype is xlsx (or
+#' xlsm) by looking at the file itself, rather than using the filename
+#' extension.
 #'
 #' @return
 #' A character vector of the names of the worksheets in the file.
@@ -14,6 +17,6 @@
 #' @examples
 #' examples <- system.file("extdata/examples.xlsx", package = "tidyxl")
 #' xlsx_sheet_names(examples)
-xlsx_sheet_names <- function(path) {
+xlsx_sheet_names <- function(path, check_filetype = TRUE) {
   utils_xlsx_sheet_files(path)$name
 }