+++ tests/testthat/test-maybe_xlsx.R
@@ -1,0 +1,11 @@
+context("maybe_xlsx")
+
+test_that("non-xlsx files are detected", {
+  expect_equal(maybe_xlsx("examples.xlsx"), TRUE)
+  expect_equal(maybe_xlsx("examples.xlsm"), TRUE)
+  expect_equal(maybe_xlsx("examples.xlsb"), TRUE) # Unfortunately xlsb look like xlsx
+  expect_equal(maybe_xlsx("examples.xls"), FALSE)
+  expect_error(xlsx_cells("examples.xlsx", check_filetype = FALSE), NA)
+  expect_error(xlsx_cells("examples.xlsb", check_filetype = FALSE), "Evaluation.*")
+  expect_error(xlsx_cells("examples.xls"), "The file format*")
+})